requirements.txt
torch==1.3.0
tensorboard==2.2.1
tqdm==4.36.1
torchvision==0.4.2
numpy==1.17.2
Pillow==7.1.2
